Change the way endpoints work
Endpoints are now a separate configuration entity from their attached
handler/interceptor/whatever.

This means that multiple endpoints can safely refer to the same
component (provided they are given different names).

This commit also changes the logic by which names are automatically
assigned: the :arachne/id is now used in preference to the handler
function name, if an :arachne/id is present.

Finally, this commit also adds support for the :any method. It is
presumed that HTTP implementors will handle this appropriately.